requirements.txt
types-redis==3.5.6
types-requests==2.25.9
types-tabulate==0.8.2
types-boto3==1.0.1

# Formatting, tests, and types.
black==21.7b0
mypy==0.910
pytest==6.2.4

# Assorted.
aiohttp~=3.7.4.0
GPUtil==1.4.0
attrs~=21.4.0
boto3==1.17.106
chardet==4.0.0
comet-ml==3.15.4
deepspeed==0.5.5
dvc[all]==2.7.3
einops~=0.4.0
fabric==2.6.0
fasttext==0.9.2
ilock==1.0.3
invoke==1.6.0
jupyterlab==3.0.16
multiprocess==0.70.12.2
msgpack~=1.0.3
nltk==3.6.2
numpy==1.21.2
openai==0.10.2
pathlib==1.0.1
plotly==5.1.0
pytz==2021.3
rpyc==5.0.1
scipy==1.8.0
sentencepiece==0.1.96
setuptools==57.4.0
tabulate==0.8.9
tqdm==4.61.2
git+https://github.com/daniel-ziegler/transformers@debertav2-fast-tokenizer-merged
typeguard==2.13.3
watchdog==2.1.6
web-pdb==1.5.6
jax==0.3.4 -f https://storage.googleapis.com/jax-releases/jax_releases.html
# want to add jax[cuda], but that doesn't work on CI because that doesn't have cuda?
# installing in pre_requirements_cpu.txt doesn't allow me to put [cuda] here
# jaxlib -f https://storage.googleapis.com/jax-releases/jax_releases.html
flax==0.4.0
seaborn==0.11.2
websockets~=10.1.0

# SQL.
psycopg2-binary==2.9.1
sqlalchemy==1.4.22

# Webapp.
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
certifi==2020.12.5
click==8.0.1
discord.py==1.7.3
flask-login==0.5.0
flask==2.0.1
flask_sqlalchemy==2.5.1
fancy_einsum>=0.0.3
gunicorn==20.0.4
idna==3.2
itsdangerous==2.0.1
redis==3.5.3
requests==2.26.0
slack_sdk==3.11.2
urllib3==1.26.6

# Sentry error monitoring.
blinker==1.4
sentry-sdk==1.4.1

# Testing.
selenium==3.141.0

git+https://github.com/google/gin-config.git@e518c4ec7755a3e5da973e894ab23cc80c6665ed#egg=gin-config

matplotlib~=3.4.2
websocket-client~=1.2.1
botocore~=1.20.106
tornado~=6.1
psutil~=5.8.0
bcrypt~=3.2.0
pyyaml~=6.0
typing_extensions~=4.1.1
ipython~=7.29.0
packaging~=21.3
pandas~=1.4.1
